Move types and funcs to better locations. Refactor blockPoller to be more generic
