+++ chain/cosmos/types.go
@@ -1,0 +1,90 @@
+package cosmos
+
+const (
+	ProposalVoteYes        = "yes"
+	ProposalVoteNo         = "no"
+	ProposalVoteNoWithVeto = "noWithVeto"
+	ProposalVoteAbstain    = "abstain"
+
+	ProposalStatusUnspecified   = "PROPOSAL_STATUS_UNSPECIFIED"
+	ProposalStatusPassed        = "PROPOSAL_STATUS_PASSED"
+	ProposalStatusFailed        = "PROPOSAL_STATUS_FAILED"
+	ProposalStatusRejected      = "PROPOSAL_STATUS_REJECTED"
+	ProposalStatusVotingPeriod  = "PROPOSAL_STATUS_VOTING_PERIOD"
+	ProposalStatusDepositPeriod = "PROPOSAL_STATUS_DEPOSIT_PERIOD"
+)
+
+// TxProposal contains chain proposal transaction details.
+type TxProposal struct {
+	// The block height.
+	Height uint64
+	// The transaction hash.
+	TxHash string
+	// Amount of gas charged to the account.
+	GasSpent int64
+
+	// Amount deposited for proposal.
+	DepositAmount string
+	// ID of proposal.
+	ProposalID string
+	// Type of proposal.
+	ProposalType string
+}
+
+// SoftwareUpgradeProposal defines the required and optional parameters for submitting a software-upgrade proposal.
+type TextProposal struct {
+	Deposit     string
+	Title       string
+	Description string
+	Expedited   bool
+}
+
+// SoftwareUpgradeProposal defines the required and optional parameters for submitting a software-upgrade proposal.
+type SoftwareUpgradeProposal struct {
+	Deposit     string
+	Title       string
+	Name        string
+	Description string
+	Height      uint64
+	Info        string // optional
+}
+
+// ProposalResponse is the proposal query response.
+type ProposalResponse struct {
+	ProposalID       string                   `json:"proposal_id"`
+	Content          ProposalContent          `json:"content"`
+	Status           string                   `json:"status"`
+	FinalTallyResult ProposalFinalTallyResult `json:"final_tally_result"`
+	SubmitTime       string                   `json:"submit_time"`
+	DepositEndTime   string                   `json:"deposit_end_time"`
+	TotalDeposit     []ProposalDeposit        `json:"total_deposit"`
+	VotingStartTime  string                   `json:"voting_start_time"`
+	VotingEndTime    string                   `json:"voting_end_time"`
+}
+
+type ProposalContent struct {
+	Type        string `json:"@type"`
+	Title       string `json:"title"`
+	Description string `json:"description"`
+}
+
+type ProposalFinalTallyResult struct {
+	Yes        string `json:"yes"`
+	Abstain    string `json:"abstain"`
+	No         string `json:"no"`
+	NoWithVeto string `json:"no_with_veto"`
+}
+
+type ProposalDeposit struct {
+	Denom  string `json:"denom"`
+	Amount string `json:"amount"`
+}
+
+type DumpContractStateResponse struct {
+	Models []ContractStateModels `json:"models"`
+}
+
+type ContractStateModels struct {
+	Key   string `json:"key"`
+	Value string `json:"value"`
+}
