+++ ibc/tx.go
@@ -32,34 +32,3 @@ func (tx Tx) Validate() error {
 	}
 	return multierr.Append(err, tx.Packet.Validate())
 }
-
-const (
-	ProposalVoteYes        = "yes"
-	ProposalVoteNo         = "no"
-	ProposalVoteNoWithVeto = "noWithVeto"
-	ProposalVoteAbstain    = "abstain"
-
-	ProposalStatusUnspecified   = "PROPOSAL_STATUS_UNSPECIFIED"
-	ProposalStatusPassed        = "PROPOSAL_STATUS_PASSED"
-	ProposalStatusFailed        = "PROPOSAL_STATUS_FAILED"
-	ProposalStatusRejected      = "PROPOSAL_STATUS_REJECTED"
-	ProposalStatusVotingPeriod  = "PROPOSAL_STATUS_VOTING_PERIOD"
-	ProposalStatusDepositPeriod = "PROPOSAL_STATUS_DEPOSIT_PERIOD"
-)
-
-// TxProposal contains chain proposal transaction details.
-type TxProposal struct {
-	// The block height.
-	Height uint64
-	// The transaction hash.
-	TxHash string
-	// Amount of gas charged to the account.
-	GasSpent int64
-
-	// Amount deposited for proposal.
-	DepositAmount string
-	// ID of proposal.
-	ProposalID string
-	// Type of proposal.
-	ProposalType string
-}
